implement GetProjectCLI
