--- src/main/kotlin/presentation/view/project/GetProjectCLI.kt
@@ -1,16 +1,0 @@
-package org.example.presentation.view.project
-
-import org.example.presentation.presenter.project.GetProjectPresenter
-import java.util.UUID
-
-class GetProjectCLI(
-    private val presenter: GetProjectPresenter
-) {
-    fun displayProject(projectId: UUID) {
-        presenter.getProjectById(projectId)
-    }
-
-    fun displayProjects() {
-        presenter.getProjects()
-    }
-}
